pyproject.toml
[tool.poetry]
name = "langchain-playground"
version = "0.1.0"
description = ""
authors = ["Okamoto Daisuke <okadaih11023@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "*"
load-dotenv = "^0.1.0"
tiktoken = "*"
langchain = "*"
langsmith = "*"
streamlit = "*"
streamlit-chat = "*"
google-api-python-client = "^2.96.0"
playwright = "^1.37.0"
beautifulsoup4 = "^4.12.2"
chroma = "^0.2.0"
chromadb = "^0.4.6"
pysqlite3-binary = "^0.5.1"
html2text = "^2020.1.16"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.265"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
poethepoet = "^0.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 200

[tool.black]
line-length = 200

[tool.coverage.xml]
output = ".pytest_cache/coverage.xml"

[tool.coverage.run]
branch = true
source = ["mymodule", "tests"]
omit = ["tests/fixtures/*"]
data_file = ".pytest_cache/.coverage"

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
]

[tool.poe.tasks]
run= "streamlit run mymodule/app.py"
lint = "ruff check mymodule"
test = "pytest"
cover = "pytest --cov=mymodule --cov-report xml"
fmt = "black mymodule --check"
build = ["fmt", "lint", "test"]